requirements.txt
click
GitPython
pyperclip
requests 